Advertise Alert Notification Services
added advertisments for a alert notification service. to do that needed to add advertisments and then callbacks for those advertiments(make sure to add onDescriptorReadRequest as its needed if you want to allow reading of characteristics). this advertiment acts as the gatt server so the peripheral which is the client will subscribe to the advertised service and revieve notifications when the values of the characteristics are changed